+++ CHANGELOG.md
@@ -26,6 +26,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * ``compare_item_collections`` is deprecated for removal in a future release.
 
+### Fixed
+
+* Properly support audio-metadata.
+
 
 
 ## [1.0.0](https://github.com/thebigmunch/google-music-utils/releases/tag/1.0.0) (2018-10-19)